requirements.txt
asgiref==3.11.1
attrs==26.1.0
black==26.3.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
coverage==7.13.5
cryptography==46.0.6
Django==6.0.3
djangorestframework==3.17.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
loguru==0.7.3
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
pycparser==3.0
Pygments==2.19.2
pyOpenSSL==26.0.0
pytest==9.0.2
pytest-cov==7.1.0
pytest-django==4.12.0
python-dotenv==1.2.2
pytokens==0.4.1
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
requests==2.33.0
rpds-py==0.30.0
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
